requirements.txt
Django==4.1.4
djangorestframework==3.14.0
argon2-cffi==21.2.0
python-dotenv==0.21.0
redis==4.4.0
PyJWT==2.6.0
kavenegar==1.1.2
Pillow==9.3.0
unicode-slugify==0.1.5
django-filter
django-extensions==3.2.1
gunicorn==20.1.0
django-cors-headers==3.13.0
pymongo==3.12.3
djongo==1.3.6
mongoengine==0.29.1
django-mongoengine==0.5.6
uvloop==0.21.0
httptools==0.6.4
coverage==6.4.4
requests
whitenoise
django-ckeditor==6.5.0
django-import-export==4.0.0
boto3
django-storages
django-rest-swagger==2.2.0
drf-yasg==1.21.7
openai==1.46.0
pandas==2.2.3
python-decouple==3.8